+++ module-apps/application-messages/ApplicationMessages.hpp
@@ -4,6 +4,7 @@
 #include <Interface/ThreadRecord.hpp>
 #include <Interface/SMSTemplateRecord.hpp>
 #include <Interface/SMSRecord.hpp>
+#include <PhoneNumber.hpp>
 
 namespace gui
 {
@@ -56,11 +57,11 @@ namespace app
         /// show dialog with big search icon and text which was used for query
         bool searchEmpty(const std::string &query = "");
         bool showSearchResults(const UTF8 &title, const UTF8 &search_text);
-        bool sendSms(const UTF8 &number, const UTF8 &body);
+        bool sendSms(const utils::PhoneNumber::View &number, const UTF8 &body);
         bool resendSms(const SMSRecord &record);
         bool newMessageOptions(const std::string &requestingWindow, gui::Text *text);
         bool showNotification(std::function<bool()> action, bool ignoreCurrentWindowOnStack = false);
-        bool handleSendSmsFromThread(const UTF8 &number, const UTF8 &body);
+        bool handleSendSmsFromThread(const utils::PhoneNumber::View &number, const UTF8 &body);
 
         // used by sms template items
         std::function<bool(std::shared_ptr<SMSTemplateRecord> templ)> templatesCallback;